+++ ChangeLog.md
@@ -28,6 +28,10 @@ Other enhancements:
 * In YAML configuration files, the `default-init-snapshot` key is introduced to
   allow a default snapshot to be specified for use with the `stack init`
   command, as if it had been specified at the command line.
+* Add flag `--haddock-executables` to Stack's `build` command (including the
+  `haddock` synonym for `build --haddock`) to enable also building Haddock
+  documentation for executables. Due to a bug in Cabal (the library), Stack will
+  ignore the flag with a warning for GHC versions before 9.4.
 
 Bug fixes: